Move errors to their own module / file
